msm: ADSPRPC: Enable remote heap with SSR support
The remote heap feature allows for the remote processor to
request memory from application subsystem for use cases where
it does not have enough memory available locally and later
release the memory when use case ends. This memory is XPU
protected on allocation and in cases of subsystem restart this
memory would need to be unlocked and freed.

Change-Id: I733915fe09715b149824a324a41fcbe8da1e3ebf
Acked-by: Himateja Reddy <hmreddy@qti.qualcomm.com>
Signed-off-by: Sathish Ambley <sathishambley@codeaurora.org>