panel: dt2w: fix display does not turn on properly
this does NOT fix the general "dt2w does not work sometimes" issue
but often dt2w works (according to logcat) but the display does not
turn on.
tbh this commit is more a (dirty..) workaround as it not really fixes
the root cause and so simply send LCD_EVENT_TOUCH_LPWG_OFF 2 times instead
of a single one.

note:
LCD_EVENT_TOUCH_LPWG_OFF = turn screen on
LCD_EVENT_TOUCH_LPWG_ON  = turn screen off

the real fix must identify why ctrl_pdata->ndx is empty when it should not
(or maybe the reason is complete elsewhere..)

Signed-off-by: steadfasterX <steadfasterX@gmail.com>