drivers: input: touch: *Correct the path for TOUCH_WAKE_COUNTER_LOG_PATH mnt partition is not used by custom roms.
	*Get tsi data only when there is a "ts" input.
I didn't like the previous way "if (!count)".
To execute the commands only when there is an input,
it is cleaner and simpler and faster.

	*Add "break" to exit the "for" loop as soon as
the "ts" input data is saved. I don't see any reason we should
keep running the commands for "count" times.

	*Remove the case where we set PARTIAL_LPWG_ON = 1.
Our power hal never had support for a partial wake up event.
It confused things. Sometimes it used to set to 1 and
the screen was dark or never turned on until we reboot
the device.

Change-Id: I22ce5515b00b73b278b0a59a944ec1cec7415292